requirements.txt
flask
pandas
scikit-learn
PyPDF2
python-docx
